pyproject.toml
[project]
name = "fgvc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.7.0",
    "huggingface-hub[cli]>=1.7.1",
    "hydra-core>=1.3.2",
    "modelscope>=1.35.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
    "bytecode>=0.17.0",
    "oss2>=2.19.1",
    "addict>=2.4.0",
    "hf-transfer>=0.1.9",
]
